+++ classes/Services.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Neat\System;
+
+use Neat\Service\Container;
+
+interface Services
+{
+    /**
+     * Register services
+     *
+     * @param Container $container
+     */
+    public function services(Container $container);
+}