+++ tests/Stub/UserModule.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Neat\System\Test\Stub;
+
+class UserModule
+{
+    /**
+     * Do stuff
+     */
+    public function stuff()
+    {
+    }
+}