+++ changelog.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2019-10-31
+### Added
+- Modules implementation.
+- Services interface.
+
 ## [0.4.0] - 2019-10-31
 ### Changed
 - Fix release.